backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
tavily-python>=0.5.0
# Agent runtime
langgraph>=0.3
langchain-core>=0.3
langchain-openai>=0.2
